+++ src/data/home.ts
@@ -9,3 +9,44 @@ export const bannerImages = [
 export const section1Text = {
   subtitle: '수만은 정밀 제조 분야의 깊은 know-how를 바탕으로, 고객 맞춤형 솔루션을 제공합니다.',
 };
+
+// //section2
+// export const section2 = {
+//   title: '산업을 움직이는 기술, 수만',
+//   description:
+//     '수만은 2차전지 생산 장비 및 신뢰성 평가 시스템을 설계/제작하는 전문 기업입니다.\n대전 R&D 센터에서 축적한 기술력과 생산 인프라를 기반으로, 고객 맞춤형 고신뢰성 솔루션을 제공합니다.',
+//   bgImage: '/images/main_section1.png',
+//   buttonLabel: 'CEO 인사말',
+// };
+
+//section3
+export const section3 = {
+  title:'정밀 부품, 모듈, 자동화 장비까지',
+  subtitle: '미래 산업에 필요한 핵심 솔루션을 제조합니다',
+  cards: [
+    {
+      title:'이차전지',
+      subtitle: 'Secondary Battery',
+      description: '고정밀 부품 / 모듈 설계',
+      img: '/images/secondary_battery.png'
+    },
+    {
+        title:'전기·전자',
+        subtitle: 'Electrical & Electronics',
+        description: '고정밀 부품 / 맞춤형 설비제작기술',
+        img: '/images/electronics.png'
+    },
+    {
+        title:'반도체',
+        subtitle: 'Semiconductor',
+        description: '고정밀 부품 / 솔루션 서비스 기술 융합',
+        img: '/images/semiconductor.png'
+    },
+    {
+        title:'자동차',
+        subtitle: 'Mobility',
+        description: '고정밀 가공기술',
+        img: '/images/mobility.png'
+    },
+  ]
+}
